requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
Django==4.1.1
django-environ==0.9.0
django-phonenumber-field==7.0.0
gunicorn==20.1.0
phonenumberslite==8.12.57
Pillow==9.2.0
psycopg2-binary==2.9.5
sqlparse==0.4.2
tk==0.1.0
